control_plane/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2-binary==2.9.9
boto3==1.34.50
python-dotenv==1.0.1
